Add Project Details Add seperate fab to worker and project details. Delete extra bottom sheet
