+++ app/src/main/java/com/ashstudios/safana/activities/SupervisorDashboard.java
@@ -3,17 +3,21 @@ package com.ashstudios.safana.activities;
 import android.content.Intent;
 import android.os.Bundle;
 
-import com.ashstudios.safana.AddEmployeesActivity;
 import com.ashstudios.safana.Fragments.BottomSheetSortFragment;
 import com.ashstudios.safana.R;
+import com.ashstudios.safana.others.Msg;
 import com.ashstudios.safana.ui.worker_details.WorkerDetailsFragment;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
-import com.google.android.material.snackbar.Snackbar;
 
 import android.view.MenuItem;
 import android.view.View;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
 import androidx.navigation.NavController;
+import androidx.navigation.NavDestination;
 import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
@@ -27,6 +31,8 @@ import androidx.appcompat.widget.Toolbar;
 
 import android.view.Menu;
 
+import java.util.List;
+
 public class SupervisorDashboard extends AppCompatActivity {
 
     private AppBarConfiguration mAppBarConfiguration;
@@ -38,29 +44,19 @@ public class SupervisorDashboard extends AppCompatActivity {
         setContentView(R.layout.activity_supervisor_dashboard);
         Toolbar toolbar = findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
-        FloatingActionButton fab = findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-//                Snackbar.make(view, "Replace with your own action", Snackbar.LENGTH_LONG)
-//                        .setAction("Action", null).show();
-                Intent generateEmployeeId = new Intent(getBaseContext(), AddEmployeesActivity.class);
-                startActivity(generateEmployeeId);
-            }
-        });
+
         DrawerLayout drawer = findViewById(R.id.drawer_layout);
-        NavigationView navigationView = findViewById(R.id.nav_view);
+        final NavigationView navigationView = findViewById(R.id.nav_view);
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         mAppBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.nav_home, R.id.nav_gallery, R.id.nav_slideshow,
+                R.id.nav_worker, R.id.nav_project_details, R.id.nav_slideshow,
                 R.id.nav_tools, R.id.nav_share, R.id.nav_send)
                 .setDrawerLayout(drawer)
                 .build();
-        NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
+        final NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment);
         NavigationUI.setupActionBarWithNavController(this, navController, mAppBarConfiguration);
         NavigationUI.setupWithNavController(navigationView, navController);
-
         workerSortBundle = new Bundle();
     }
