+++ app/src/main/java/com/ashstudios/safana/ui/project__details/ProjectDetailsViewModel.java
@@ -1,14 +1,14 @@
-package com.ashstudios.safana.ui.gallery;
+package com.ashstudios.safana.ui.project__details;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class GalleryViewModel extends ViewModel {
+public class ProjectDetailsViewModel extends ViewModel {
 
     private MutableLiveData<String> mText;
 
-    public GalleryViewModel() {
+    public ProjectDetailsViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is gallery fragment");
     }
