+++ app/src/main/java/com/ashstudios/safana/ui/worker_details/WorkerDetailsFragment.java
@@ -1,8 +1,10 @@
 package com.ashstudios.safana.ui.worker_details;
 
 import android.content.Context;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
@@ -16,8 +18,10 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.ashstudios.safana.R;
+import com.ashstudios.safana.activities.AddEmployeesActivity;
 import com.ashstudios.safana.activities.SupervisorDashboard;
 import com.ashstudios.safana.adapters.WorkerRVAdapter;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class WorkerDetailsFragment extends Fragment {
 
@@ -28,12 +32,22 @@ public class WorkerDetailsFragment extends Fragment {
                              ViewGroup container, Bundle savedInstanceState) {
         workerDetailsViewModel = ViewModelProviders.of(this).get(WorkerDetailsViewModel.class);
         View root = inflater.inflate(R.layout.fragment_worker_details, container, false);
+        FloatingActionButton fab = root.findViewById(R.id.fab);
 
         recyclerView = root.findViewById(R.id.rc_worker_details);
         WorkerRVAdapter workerRVAdapter = new WorkerRVAdapter(workerDetailsViewModel,getContext());
         recyclerView.setNestedScrollingEnabled(false);
         recyclerView.setAdapter(workerRVAdapter);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+
+        fab.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                    Intent generateEmployeeId = new Intent(getContext(), AddEmployeesActivity.class);
+                    startActivity(generateEmployeeId);
+            }
+        });
+
 
         return root;
     }
