+++ app/src/main/res/layout/app_bar_supervisor_dashboard.xml
@@ -21,16 +21,6 @@
 
     </com.google.android.material.appbar.AppBarLayout>
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/fab"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom|end"
-        android:layout_margin="@dimen/fab_margin"
-        app:borderWidth="0dp"
-        app:srcCompat="@drawable/ic_add_white_24dp"
-        android:backgroundTint="@color/dark_gray"/>
-
     <include layout="@layout/content_supervisor_dashboard" />
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
