+++ app/src/main/res/menu/activity_supervisor_dashboard_drawer.xml
@@ -5,13 +5,13 @@
 
     <group android:checkableBehavior="single">
         <item
-            android:id="@+id/nav_home"
+            android:id="@+id/nav_worker"
             android:icon="@drawable/ic_menu_camera"
             android:title="@string/menu_worker_details" />
         <item
-            android:id="@+id/nav_gallery"
+            android:id="@+id/nav_project_details"
             android:icon="@drawable/ic_menu_gallery"
-            android:title="@string/menu_gallery" />
+            android:title="@string/menu_project_details" />
         <item
             android:id="@+id/nav_slideshow"
             android:icon="@drawable/ic_menu_slideshow"
