+++ app/src/main/res/values/strings.xml
@@ -23,4 +23,5 @@
 
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="menu_project_details">Project details</string>
 </resources>
